pyproject.toml
[tool.poetry]
name = "langchain-hackathon"
version = "0.1.0"
description = ""
authors = ["Lars Petter Madsstuen <100695195+larspettermadsstuen@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "langchain_hackathon"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.283"
openai = "^0.28.0"
datamodel-code-generator = "^0.21.5"
pydantic-yaml = "^1.1.1"
pydantic = "^2.3.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
